tests: Move job.go under libnet folder
