+++ tests/libnet/job/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["job.go"],
+    importpath = "kubevirt.io/kubevirt/tests/libnet/job",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//tests:go_default_library",
+        "//tests/framework/kubevirt:go_default_library",
+        "//vendor/k8s.io/api/batch/v1:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/wait:go_default_library",
+    ],
+)
